refactor: [M3-6331] - MUI v5 - Components > EnhancedNumberInput
