+++ packages/manager/src/components/EnhancedNumberInput/EnhancedNumberInput.stories.mdx
@@ -1,6 +1,6 @@
 import Typography from 'src/components/core/Typography';
 import Paper from 'src/components/core/Paper';
-import { EnhancedNumberInput } from './EnhancedNumberInput';
+import { EnhancedNumberInput } from 'src/components/EnhancedNumberInput/EnhancedNumberInput';
 import { Canvas, Meta, Story, ArgsTable } from '@storybook/addon-docs';
 import * as React from 'react';
