+++ packages/manager/src/components/EnhancedNumberInput/EnhancedNumberInput.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { wrapWithTheme } from 'src/utilities/testHelpers';
 
-import EnhancedNumberInput from './EnhancedNumberInput';
+import { EnhancedNumberInput } from 'src/components/EnhancedNumberInput/EnhancedNumberInput';
 
 const setValue = jest.fn();
