--- packages/manager/src/components/EnhancedNumberInput/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './EnhancedNumberInput';
